server/eth: Start ETH backend if at least one provider is up to date
Previously, the ETH backend would not start if any of the providers
could not connect or was outdated. Now, if at least one of the
providers is able to connect and the header is recent, the backend
will start. Outdated providers are retried if originally working
providers start failing. Providers which falied to connect will not be
retried.

If at some point all providers become outdated, outdated providers will
be temporarily allowed for 2 minutes so that the backend doesn't stop
serving requests.