client/{mm,core}: Market maker balance segregation
New configurations are added to the market maker config specifying what
part of the wallet's balance each market maket will have control over.
Using this config, a `balanceHandler` object is created, which tracks
each of the orders made by the market makers, decreasing and increasing
their balances as needed over the lifecycle of the order. The
`balanceHandler` is not called directly by the market makers, instead a
`coreWithSegregatedBalance` is created which wraps the core methods so
that from the point of view of the bot, they behave as if the entire
balance of the wallet is only what has been allocated to that bot by
the `balanceHandler`.