client/mm: Simple arb bot auto-rebalancing
This implements auto-rebalancing in the simple arbitrage strategy. This is
an optional functionality that allows the user to specify a minimum
balance of each asset that should be on both the DEX and the CEX. If the
balance dips below this amount, then either a deposit or withdrawal will
be done to have an equal amount of the asset on both exchanges.