client/mm: Token asset balance management
This updates the market making balance management system to handle token
assets.

A new `asset.TokenCoin` interface is added that provides a `ParentAssetValue`
function that returns amount of the parent asset that is reserved for fees.
This allows a `ParentAssetLockedAmt` field to be added to `core.Order`,
which allows the market making code to know how much of the parent asset
was reserved when making a trade.