client/{asset,core,rpcserver}: Add WalletTransaction
- This adds a WalletTransaction function to the WalletHistorian interface.
It returns information about a transaction which the wallet made, or one
in which the wallet received funds. The TransactionConfirmations function
is removed, as it will no longer be needed once all wallets support
WalletTransaction.

- A `PartOfActiveBalance` field is added to the `WalletTransaction` type.
This is required for market making to know when it is safe to consider the
funds from an incoming transaction available for use.

- A `ParentAssetLockedAmt` field is added to the `core.Order` type. This is
needed in order to know the amount of token's fee asset is locked for an
order.