mm: Add BotProblems to BotStatus
This PR adds a BotProblems struct to BotStatus that reports issues that
bots are facing. Some problems such as wallet sync and connectivity are
checked preemptively while others are reported when errors arise during
determining what orders to place and when actually placing them. The
preemptive checks are especially important for the Arb bot which
additionally confirms that the trading limits allow for an additional
trade. Balance deficiencies are also reported in BotProblems.