mm: Internal transfers
This diff updates the market maker to attempt to reallocate available funds
to bots before doing a deposit or a withdrawal. For example, if a deposit
is required, and the user has additional available funds on the CEX that are
not allocated to any bots, funds will be allocated to the bot on the CEX
and unallocated from the bot on the wallet to simulate a deposit. Bots may
be configured to only attempt these "internal transfers", to attempt both
internal and external transfers, or to do no rebalancing at all.

# Conflicts:
#	client/mm/mm_arb_market_maker.go
#	client/mm/mm_simple_arb.go
#	client/webserver/locales/en-us.go
#	client/webserver/site/src/js/mm.ts